# public/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Space-Stylish E-commerce API
  description: |
    Space-Stylish E-commerce Website API Documentation

    This is a complete e-commerce platform API that provides the following features:
    - User registration, login and profile management
    - Product browsing, search and detailed information
    - Shopping cart and order management
    - Marketing campaign management
    - Sales reports and analytics
    - Admin functions

    ## Authentication
    Most APIs require JWT token authentication. Please include in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Rate Limiting
    - General APIs: 60 requests per second
    - API endpoints: 30 requests per second
    - Admin endpoints: 10 requests per second
    - Health check and static resources are excluded

    ## Response Format
    All APIs return a standardized response format:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message",
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ## Environment Variables
    - `API_VERSION`: API version (default: v1)
    - `SERVER_PORT`: Server port
    - `JWT_SIGN_SECRET`: JWT signing secret
    - `PARTNER_KEY`: TapPay partner key
    - `MERCHANT_ID`: TapPay merchant ID
    - `APP_ID`: Facebook App ID
    - `APP_SECRET`: Facebook App Secret

  version: 2.0.0
  contact:
    name: Space-Stylish Team
    email: support@space-stylish.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /admin/create_role:
    post:
      tags:
        - Admin
      summary: Create role
      description: Create a new user role (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - access
              properties:
                role:
                  type: string
                  description: Role name
                access:
                  type: string
                  description: Role access permissions
      responses:
        "200":
          description: Role created successfully
          content:
            application/json:
              schema:
                type: string
                example: "New Role Successfully Created"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to create role"
  /admin/assign_role:
    post:
      tags:
        - Admin
      summary: Assign role to user
      description: Assign a role to a specific user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - roleId
              properties:
                userId:
                  type: integer
                  description: User ID
                roleId:
                  type: integer
                  description: Role ID
      responses:
        "200":
          description: Role assigned successfully
          content:
            application/json:
              schema:
                type: string
                example: "New Role Successfully Assigned"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to assign role"
  /admin/create_product:
    post:
      tags:
        - Admin
      summary: Create product with images
      description: Create a new product with image uploads (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - category
                - title
                - price
              properties:
                main_image:
                  type: string
                  format: binary
                  description: Main product image
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Additional product images
                category:
                  type: string
                  enum: [women, men, accessories]
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                texture:
                  type: string
                wash:
                  type: string
                place:
                  type: string
                note:
                  type: string
                story:
                  type: string
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      color_name:
                        type: string
                      color_code:
                        type: string
                      size:
                        type: string
                      stock:
                        type: integer
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                type: string
                example: "success"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to create product"
  /admin/create_campaign:
    post:
      tags:
        - Admin
      summary: Create marketing campaign
      description: Create a new marketing campaign (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - product_id
                - story
              properties:
                picture:
                  type: string
                  format: binary
                  description: Campaign image
                product_id:
                  type: integer
                  description: Associated product ID
                story:
                  type: string
                  description: Campaign story
      responses:
        "200":
          description: Campaign created successfully
          content:
            application/json:
              schema:
                type: string
                example: "New Campaign Successfully Created"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: string
                example: "chosen item does not exist"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to create campaign"
  /admin/get_orders:
    get:
      tags:
        - Admin
      summary: Get orders from external system
      description: Get orders from Azole system (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Orders data from external system
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
                example: "Failed to get orders"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...